refactor Printer
