+++ Dzmitry-Monich/2/checker/lib/checker/printer.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'printer/response_printer'
+
 module Checker
   class Printer
     def initialize(responses)
@@ -8,7 +10,7 @@ module Checker
     end
 
     def call
-      urls = prepare_urls
+      urls = ResponsePrinter.print(responses)
       total = prepare_total
       puts [urls, total].join("\n\n")
     end
@@ -17,30 +19,11 @@ module Checker
 
     attr_reader :responses, :total
 
-    def renderers
-      [
-        {
-          check: ->(status) { %i[success failed].include?(status) },
-          fn: ->(res) { "#{res.url} - #{res.response.status} (#{t_format(res.interval)}ms)" }
-        },
-        {
-          check: ->(status) { status == :errored },
-          fn: ->(res) { "#{res.url} - ERROR (#{res.message})" }
-        }
-      ]
-    end
-
     def total_result
       @total_result ||= responses.each_with_object(total) do |res, acc|
         acc[res.status] += 1
         acc[:total] += 1
       end
-    end
-
-    def prepare_urls
-      responses.map do |res|
-        renderers.find { |r| r.fetch(:check).call(res.status) }.fetch(:fn).call(res)
-      end.join("\n")
     end
 
     def prepare_total
@@ -50,10 +33,6 @@ module Checker
         "Failed: #{total_result[:failed]}",
         "Errored: #{total_result[:errored]}"
       ].join(', ')
-    end
-
-    def t_format(time)
-      (time * 1000).round
     end
   end
 end
